Refactor: modalIsOpenState 이동
